box ddp is working but does not yet consider collisions with road objects and does not return the grads

+++ algos/box_ddp/lqr_step.py
@@ -209,7 +209,8 @@ def LQRStep(n_state,
                         lb[I] = lb_limit if isinstance(lb_limit, float) else lb_limit[I]
                         I = ub > ub_limit
                         ub[I] = ub_limit if isinstance(lb_limit, float) else ub_limit[I]
-                    new_ut = util.eclamp(new_ut, lb, ub)
+                    # TODO(eugenevinitsky) why do we need to do this?
+                    new_ut = util.eclamp(new_ut, lb.double(), ub.double())
                 new_u.append(new_ut)
 
                 new_xut = torch.cat((new_xt, new_ut), dim=1)
